Add divs to index.html and style styles.css to match required design
Added a container to group all the cards.
Split header, content and footer in three diferent divs. Content has flex-grow = 1 to occupy all the space.
Applied flex-wrap to cards, as well as fixed width so it fits 3 on a normal sized screen and 2 on a smaller one (in the same row).